fix and add add return data

+++ deletegroup.php
@@ -4,6 +4,16 @@ header('Content-Type: application/json');
 include "dbconnect.php";
 include "finduserid.php";
 
+function isAdmin($con,$groupid,$userid){
+	$sqlisadmin = "SELECT * FROM `has_user` WHERE `user_id` = '$userid' AND `role_id` = 1 and `group_id` = '$groupid'";
+	$queryisadmin = $con->query($sqlisadmin);
+	if($queryisadmin->num_rows>0){
+		return true;
+	}else{
+		return false;
+	}
+}
+
 $sessionid = $con->real_escape_string($_GET['sessionid']);
 $groupid = $con->real_escape_string($_GET['groupid']);
 $type = $con->real_escape_string($_GET['type']);
@@ -13,50 +23,51 @@ $userid = finduserid($sessionid,$con,$type);
 
 if($userid != 0){
 
-	//check ว่า groupid ของเรามี parent_id อะไร ถ้าไม่มีใส่เป็น 0
-	$sqlcheckparentid = "SELECT `parent_id` FROM `group` WHERE `group_id` = '$groupid'";
-	$queryparentid = $con->query($sqlcheckparentid);
-	$row = $queryparentid->fetch_assoc();
-	$parentid = $row['parent_id'];
-	if($parentid != ""){
-			//check ว่ามีกลุ่มไหนมี parent_id เป็น groupid ที่ส่งมาบ้าง
-			$groupidtochange = array();
-			$sqlgetgroup = "SELECT * FROM `group` WHERE `parent_id` = '$groupid'";
-			$querycheck = $con->query($sqlgetgroup);
-			if($querycheck -> num_rows > 0){
-				while ($row2 = $querycheck->fetch_assoc()) {
-					//เปลี่ยน parentid
-					array_push($groupidtochange,$row2['group_id']);
-				}
-				//เปลี่ยน parentid ของกลุ่มนั้นให้เป็นตาม $parentid
-				$str_groupidtochange = implode(",",$groupidtochange);
-				$sqlchangeparent = "UPDATE `workingalert`.`group` SET `parent_id` = '$parentid' WHERE `group`.`group_id` in ($str_groupidtochange);";
-				//echo $sqlchangeparent;
-				if($con->query($sqlchangeparent)){
-					//change success
-					if($con->query("DELETE FROM `workingalert`.`group` WHERE `group`.`group_id` = '$groupid'")==true){
+	if( isAdmin($con,$groupid,$userid) ) {
+
+		//check ว่า groupid ของเรามี parent_id อะไร ถ้าไม่มีใส่เป็น 0
+		$sqlcheckparentid = "SELECT `parent_id` FROM `group` WHERE `group_id` = '$groupid'";
+		$queryparentid = $con->query($sqlcheckparentid);
+		$row = $queryparentid->fetch_assoc();
+		$parentid = $row['parent_id'];
+		if($parentid != ""){
+				//check ว่ามีกลุ่มไหนมี parent_id เป็น groupid ที่ส่งมาบ้าง
+				$groupidtochange = array();
+				$sqlgetgroup = "SELECT * FROM `group` WHERE `parent_id` = '$groupid'";
+				$querycheck = $con->query($sqlgetgroup);
+				if($querycheck -> num_rows > 0){
+					while ($row2 = $querycheck->fetch_assoc()) {
+						//เปลี่ยน parentid
+						array_push($groupidtochange,$row2['group_id']);
+					}
+					//เปลี่ยน parentid ของกลุ่มนั้นให้เป็นตาม $parentid
+					$str_groupidtochange = implode(",",$groupidtochange);
+					$sqlchangeparent = "UPDATE `workingalert`.`group` SET `parent_id` = '$parentid' WHERE `group`.`group_id` in ($str_groupidtochange);";
+					//echo $sqlchangeparent;
+					if($con->query($sqlchangeparent)){
+						//change success
+						if($con->query("DELETE FROM `workingalert`.`group` WHERE `group`.`group_id` = '$groupid'")==true){
+							$response = array("status"=>"success","description"=>"delete success");
+						}else{
+							$response = array("status"=>"failed","description"=>"delete row not complete");
+						}
+					}else{
+						$response = array("status"=>"failed","description"=>"change child failed");
+					}
+				}else{
+					//ถ้าไม่มีก็ลบกลุ่มได้เลย
+					if($con->query("DELETE FROM `workingalert`.`group` WHERE `group`.`group_id` = '$groupid'")){
 						$response = array("status"=>"success","description"=>"delete success");
 					}else{
 						$response = array("status"=>"failed","description"=>"delete row not complete");
 					}
-				}else{
-					$response = array("status"=>"failed","description"=>"change child failed");
 				}
-			}else{
-				//ถ้าไม่มีก็ลบกลุ่มได้เลย
-				if($con->query("DELETE FROM `workingalert`.`group` WHERE `group`.`group_id` = '$groupid'")){
-					$response = array("status"=>"success","description"=>"delete success");
-				}else{
-					$response = array("status"=>"failed","description"=>"delete row not complete");
-				}
+			}else {
+				$response = array("status"=>"failed","description"=>"not fount new parent id");
 			}
-
-
 	}else {
-		$response = array("status"=>"failed","description"=>"not fount new parent id");
+		$response = array("status"=>"failed","description"=>"Delete fail, because you are not admin of this group");
 	}
-
-
 }else{
 	$response = array("status"=>"failed","description"=>"not have permission");
 }
